Refactor create challenge by creating helper function to improve code readability
