+++ routes/api/play_hunts.js
@@ -5,15 +5,12 @@ const mongoose = require("mongoose");
 const validatePlayHuntInput = require("../../validation/play_hunt");
 const PlayHunt = require("../../models/PlayHunt");
 const Hunt = require("../../models/Hunt")
-var AWS = require("aws-sdk");
 var multer = require('multer');
 var upload = multer({ dest: 'uploads/' });
-const keys = require('../../config/keys');
-var fs = require('file-system');
-var fsy = require("fs");
+
 var resemble = require('resemblejs');
 var http = require("http");
-const hunt = require("../../validation/hunt");
+const PlayHuntHelper = require("../../helper/play_hunts_helper");
 
 router.get("/test", (req, res) => res.json({ msg: "This is the play hunts route" }));
 
@@ -48,88 +45,38 @@ router.post("/", [passport.authenticate('jwt', { session: false }), upload.array
     // fetch hunt based on hunt_id
     Hunt.findById(req.body.hunt_id)
         .then(playHunt => {
-            // validating based on hunt images
+            // validate req images
             const { errors, isValid } = validatePlayHuntInput(req, playHunt);
             if (!isValid) return res.status(400).json(errors);
+            debugger
+            //create custom promise to make code more readable
+            PlayHuntHelper.uploadToAWS(req, imageAwsPaths=>{
 
-            let s3bucket = new AWS.S3({
-                accessKeyId: keys.AWS_ACCESS_KEY_ID,
-                secretAccessKey: keys.AWS_SECRET_ACCESS_KEY,
-                region: keys.AWS_REGION
-            });
+                //images uploaded to aws and imageAwsPaths is path to those images
+                let scoreSum = 0, numFiles = imageAwsPaths.length;
 
-            const numFiles = req.files.length;
-            const imageAwsPath = new Array(numFiles); 
-            imageAwsPath.fill(null);
-            
+                //iterating through images sending pairs to rememeble to get the match score
+                for (let i = 0; i < numFiles; i++) {
+                    resemble(imageAwsPaths[i])
+                        .compareTo(playHunt.photo_collection[i])
+                        .scaleToSameSize()
+                        .ignoreAntialiasing()
+                        .onComplete(function (data) {
+                            scoreSum += 100 - data.misMatchPercentage;
+                            if (numFiles - i === 1) {
+                               //we have result of all images summed up in scoreSum
 
-
-            //iterating through the files submitted in the request
-            req.files.map((item, idx) => {
-                //setting params for aws
-                var params = {
-                    Bucket: keys.AWS_BUCKET_NAME,
-                    Key: item.originalname,
-                    Body: fs.createReadStream(item.path),
-                    ACL: 'public-read'
-                }
-                // uploading image to aws
-                s3bucket.upload(params, function (err, data) {
-                    if (err) {
-                        // error 
-                        res.json({ "error": true, "Message": err });
-                    } else {
-                        // success
-
-                        imageAwsPath[idx] = data.Location;
-                        if (numFiles === imageAwsPath.length && !imageAwsPath.includes(null)) {
-                            let ele = 0;
-                            for (let i = 0; i < numFiles; i++) {
-                                resemble(imageAwsPath[i])
-                                    .compareTo(playHunt.photo_collection[i])
-                                    .scaleToSameSize()
-                                    .ignoreAntialiasing()
-                                    .onComplete(function (data) {
-                                        ele += 100 - data.misMatchPercentage;
-                                        if (numFiles - i === 1) {
-                                            const playHunt = new PlayHunt({
-                                                user: req.user.id,
-                                                hunt_id: req.body.hunt_id,
-                                                timestamps: req.body.timestamps,
-                                                score: ele / numFiles,
-                                                images: imageAwsPath
-                                            })                                            
-                                            playHunt.save()
-                                            Hunt.findById(req.body.hunt_id)
-                                            .then((hunt) => {
-                                                if (hunt.winner.score < playHunt.score) {
-                                                    Hunt.updateOne(
-                                                        {_id: playHunt.hunt_id},
-                                                        {"winner": {id: playHunt.user, score: playHunt.score}
-                                                        }).catch(err => {
-                                                         return  res.json(err)
-                                                       })
-                                                }
-                                                //building a custom response
-                                                const customResponse = {
-                                                    hunt_id: hunt._id,
-                                                    score: playHunt.score,
-                                                    hunt_name: hunt.title
-                                                }
-                                                return res.json(customResponse);
-                                            })
-                                        }
-                                        
-                                    })
+                                const avgScore = scoreSum / numFiles;
+                                PlayHuntHelper.completeChallenge(avgScore,imageAwsPaths, req, (customResponse)=>{
+                                    return res.json(customResponse);
+                                });
                             }
-                        }
-                    }
-                })
-            });
+                        });
+                }
+            })
 
         })
-
-
 })
+
 
 module.exports = router;
